backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
scipy
sqlalchemy-utils
psycopg2-binary
python-dotenv
openpyxl